Code Cleanup

+++ lttm.py
@@ -1,9 +1,13 @@
+import os, time
+
+os.environ['PYGAME_HIDE_SUPPORT_PROMPT'] = "hide"
+
 from gtts import gTTS
-from pygame import mixer #Playing sound
+from pygame import mixer
 
-import os
-
+#TODO Adicionar
 language = 'pt-br'
+mixer.init(devicename='CABLE Input (VB-Audio Virtual Cable)')
 
 
 def create_audio(text):
@@ -12,28 +16,30 @@ def create_audio(text):
 
 
 def play_audio(text):
-    mixer.init(devicename='CABLE Input (VB-Audio Virtual Cable)')
     mixer.music.unload()
     create_audio(text)
-    mixer.music.load("text.mp3") #Load the mp3
+    mixer.music.load("text.mp3")
     mixer.music.play()
 
+
 def clear():
-    os.remove('text.mp3')
-    # os.unlink()
+    mixer.music.unload()
+    mixer.quit()
+    if os.path.isfile('text.mp3'):
+        os.remove('text.mp3')
+
 
 def loop():
     while True:
-        text = input("Texto: ")
+        text = input("Text To Mic: ")
 
         if text == 'exit now':
-            print("Saindo...")
+            print("Exiting...")
             clear()
-            input('')
+            time.sleep(2)
             break
 
         play_audio(text)
-
